+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.3.2] - Unreleased
 
+### Changed
+
+- Treat generated pg_tools setup as opt-in for apps that use structure.sql or PostgreSQL CLI workflows.
+
 ## [0.3.1] - 2026-05-13
 
 ### Added